HTCONDOR-2357 Don't kill processes we've already reaped,
we might ending up killing something we didn't intent to.

+++ src/condor_daemon_core.V6/daemon_core.cpp
@@ -8624,6 +8624,9 @@ DaemonCore::Kill_Thread(int tid)
 	 */
 	return 1;
 #else
+	if (ProcessExitedButNotReaped(tid)) {
+		return true;
+	}
 	priv_state priv = set_root_priv();
 	int status = kill(tid, SIGKILL);
 	set_priv(priv);
